+++ jest.config.js
@@ -1,5 +1,6 @@
 export default {
   "testMatch": [
     "**/tests/**/*.js"
-  ]
+  ],
+  "transform": {}
 }