+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this tool will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0]
+### Added
+- updates checking
+- new tests
+- `-e, --exclude-all` option
+
+### Changed
+- bug: saving "nothing" instead of deleting. fixed.
+
 ## [2.0.0]
 ### Changed
 - bug: saving "nothing" instead of deleting. fixed.